requirements.txt
langgraph>=0.0.10
arxiv==2.0.0
langchain>=0.1.0
langchain-core>=0.1.10
langchain-groq>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.4.22
beautifulsoup4>=4.12.0
pypdf>=4.0.0 